arion/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2

plotly==5.17.0
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.1
xgboost==2.0.0
requests==2.31.0
yfinance==0.2.28
vaderSentiment==3.3.2
python-dotenv==1.0.0
ta==0.11.0

pydantic==1.10.12
